requirements.txt
Flask==3.0.0
redis==5.0.1
mysql-connector-python==8.2.0
requests==2.31.0
cachetools==5.3.2
hiredis==2.3.2
eventlet==0.35.0
gunicorn==21.2.0
gevent==23.9.1
